Log errors returned by OpenPlatform
Currently we do not provide any feedback to callers if a response
contains an error and no other signal is provided.

We do not have a proper way to communicate errors relating to a
single search to the client so the current behavior of returning no
new results make sense. However shallowing exceptions entirely makes
it impossible to know if anything has gone wrong.

This PR adds some logging to ensure that errors are at least picked
up somewhere.

Update tests accordingly.