Do not ignore scripts when installing the package
First of all PHPStan extension installer requires it to work. Second
it changes the behavior of Composer in general and developers are
unlikely to run with this option locally making things hard to debug.

Never quiet composer install as it makes it harder to figure out what
is going on.