Fix handling of query added multiple times
Our current model is based on the query being unique for the user.
However if a query is added multiple times the service will return
an error due to an unique constraint violation.

We expected this to be handled by using updateOrInsert in combination
with our unique index. However the query builder will actually query
for the existence of our attributes (first paramater) and based on
that determine if attributes *and* values (2nd parameter) should be
inserted or updated.

Consequently we move elements which are not part of the unique index
from attributes to values.

We also add a test case which covers this scenario.